+++ src/app/(home)/components/scopeCarousel/ScopeCarousel.tsx
@@ -1,11 +1,12 @@
 'use client';
 
-import React, { useState, useEffect } from 'react'; // useState는 canScroll 상태에 사용
+import { useState, useEffect, useCallback } from 'react';
 import * as styles from './scopeCarousel.css';
 import Title from '@/common/components/title/Title';
-import ScopeCarouselItem from './ScopeCarouselItem';
-import ScopeCarouselControls from './ScopeCarouselControls';
-import { SCOPE_CAROUSEL_CATEGORIES, CAROUSEL_OPTIONS } from './constants';
+import ScopeCarouselItem from './scopeCarouselItem/ScopeCarouselItem';
+import ScopeCarouselControls from './scopeCarouselControls/ScopeCarouselControls';
+import { CAROUSEL_OPTIONS } from './constants';
+import { SCOPE_CAROUSEL_DATA } from '@/mocks/data/home';
 import useEmblaCarousel from 'embla-carousel-react';
 
 const ScopeCarousel = () => {
@@ -31,8 +32,8 @@ const ScopeCarousel = () => {
     };
   }, [emblaApi]);
 
-  const scrollPrev = React.useCallback(() => emblaApi?.scrollPrev(), [emblaApi]);
-  const scrollNext = React.useCallback(() => emblaApi?.scrollNext(), [emblaApi]);
+  const scrollPrev = useCallback(() => emblaApi?.scrollPrev(), [emblaApi]);
+  const scrollNext = useCallback(() => emblaApi?.scrollNext(), [emblaApi]);
 
   return (
     <div className={styles.container}>
@@ -48,7 +49,7 @@ const ScopeCarousel = () => {
 
       <div className={styles.viewport} ref={emblaRef}>
         <div className={styles.emblaContainer}>
-          {SCOPE_CAROUSEL_CATEGORIES.map((category) => (
+          {SCOPE_CAROUSEL_DATA.map((category) => (
             <div key={category.id} className={styles.slide}>
               <ScopeCarouselItem item={category.articles.progressive} ideology="progressive" />
               <ScopeCarouselItem item={category.articles.conservative} ideology="conservative" />
