--- src/app/(home)/components/scopeCarousel/index.ts
@@ -1,4 +1,0 @@
-export { default as ScopeCarousel } from './ScopeCarousel';
-export { default as ScopeCarouselItem } from './ScopeCarouselItem';
-export { default as ScopeCarouselControls } from './ScopeCarouselControls';
-export type { IdeologyType } from './types';
